Source1/requirements.txt
python-json-logger
pyjwt
python-dotenv